Boolean -> byte

+++ mus2mid.h
@@ -22,6 +22,6 @@
 
 #include <stdio.h>
 
-int mus2mid(FILE *musinput, FILE *midioutput);
+byte mus2mid(FILE *musinput, FILE *midioutput);
 
 #endif /* #ifndef MUS2MID_H */
